Basic ScrollPositionFlow widget.
From here should be broken out into the various subtly related classes and more
fundamental components. For instance, a widget that provides offset callbacks so
that EntranceAnimation widgets can be created, and also more friendly classes
that take Tweens and Curves etc for convenience.

This behavior is also useful for more than just scroll effects, so this class
should be decoupled from scroll if possible (and maybe even made its own
library? Probably not but...)