Docs & reasonably feature rich implementations of Flow style widgets.
Flow widgets being EntranceAnimationFlow and ScrollPositionFlow.

Docs not yet added to README, nor are example videos made.

Also likely want to add an EntrancePolicy support to
EntranceAnimationFlow.

May also want to support translations done to child descendents of these
Flow widgets, so that the position of one box can drive an animation on
one of its children instead of itself.

And of course, may want to support rebuilds for cases where that's
required, even though it may open the door to some pathological edge
cases.